Updated tests and methods for multiple classes
The tests and methods for several classes, including ShapeCrawler, SlideMaster, and Shape objects have been updated. Various features such as getting the width of a shape, returning a shape Geometry type, and returning the shape xPath were implemented. Also, some improvements were made to handle ShapeCollection operations, including additions and removal of slides, counting slides, and others. The updates also provide more accurate handling of slides with charts, tables, and various geometric shapes.
