+++ test/ShapeCrawler.Tests.Unit/ShapeTests.cs
@@ -513,4 +513,58 @@ public class ShapeTests : SCTest
         // Assert
         x.Should().Be(53);
     }
+    
+    [Test]
+    [TestCase("050_title-placeholder.pptx", 1, 2, 777)]
+    [TestCase("051_title-placeholder.pptx", 1, 3074, 864)]
+    public void Width_returns_width_of_Title_placeholder(
+        string filename, 
+        int slideNumber, 
+        int shapeId,
+        int expectedWidth)
+    {
+        // Arrange
+        var pres = new Presentation(StreamOf(filename));
+        var shape = pres.Slides[slideNumber - 1].Shapes.GetById<IShape>(shapeId);
+
+        // Act
+        var shapeWidth = shape.Width;
+
+        // Assert
+        shapeWidth.Should().Be(expectedWidth);
+    }
+    
+    [Test]
+    [SlideShape("006_1 slides.pptx", 1, "Shape 2", 149)]
+    [SlideShape( "009_table.pptx", 2, "Object 3", 39)]
+    [SlideShape( "autoshape-grouping.pptx", 1, "Group 2", 108)]
+    public void Height_returns_shape_height_in_pixels(IShape shape, int expectedHeight)
+    {
+        // Act
+        var height = shape.Height;
+
+        // Assert
+        height.Should().Be(expectedHeight);
+    }
+    
+    [Test]
+    [SlideShape("021.pptx", 4, 2, Geometry.Rectangle)]
+    [SlideShape("021.pptx", 4, 3, Geometry.Ellipse)]
+    public void GeometryType_returns_shape_geometry_type(IShape shape, Geometry expectedGeometryType)
+    {
+        // Assert
+        shape.GeometryType.Should().Be(expectedGeometryType);
+    }
+    
+    [Test]
+    [SlideShape("054_get_shape_xpath.pptx", 1, "Title 1", "/p:sld[1]/p:cSld[1]/p:spTree[1]/p:sp[1]")]
+    [SlideShape("054_get_shape_xpath.pptx", 1, "SubTitle 2", "/p:sld[1]/p:cSld[1]/p:spTree[1]/p:sp[2]")]
+    public void SDKXPath_returns_shape_xpath(IShape shape, string expectedXPath)
+    {
+        // Act
+        var shapeXPath = shape.SDKXPath;
+
+        // Assert
+        shapeXPath.Should().Be(expectedXPath);
+    }
 }
